Match actions, types, reducer and sagas
