+++ src/Redux/MainPage/MainSagas.js
@@ -21,15 +21,6 @@ function* getListOfDisciplines() {
   }
 }
 
-function* sortMatches({payload}) {
-  try {
-    const sortedMatches = yield api.fetchMatchesByDiscipline(payload)
-    yield put(actionCreators.sortMatchesSuccess(sortedMatches))
-  } catch (error) {
-    console.log(error)
-  }
-}
-
 function* onGetAnnounces() {
   yield takeLatest(actionTypes.GET_ANNOUNCES, getAnnounces)
 }
@@ -38,14 +29,9 @@ function* onGetListOfDisciplines() {
   yield takeLatest(actionTypes.GET_LIST_OF_DISCIPLINES, getListOfDisciplines)
 }
 
-function* onSortMatchesByDiscipline() {
-  yield takeLatest(actionTypes.SORT_MATCHES, sortMatches)
-}
-
 export default function* mainSagas() {
   yield all([
     call(onGetAnnounces),
-    call(onGetListOfDisciplines),
-    call(onSortMatchesByDiscipline),
+    call(onGetListOfDisciplines)
   ])
 }
