+++ src/Redux/Store/RootReducer.ts
@@ -2,6 +2,7 @@ import {combineReducers} from 'redux';
 import mainReducer from '../MainPage/MainReducer';
 import storage from 'redux-persist/lib/storage'
 import { persistReducer } from 'redux-persist';
+import matchReducer from '../Match/MatchReduces';
 
 const rootPersistConfig = {
   key: 'root',
@@ -13,8 +14,14 @@ const mainPersistConfig = {
   storage: storage
 }
 
+const matchPersistConfig = {
+  key: 'match',
+  storage: storage
+}
+
 const rootReducer = combineReducers({
-  main: persistReducer(mainPersistConfig, mainReducer)
+  main: persistReducer(mainPersistConfig, mainReducer),
+  match: persistReducer(matchPersistConfig, matchReducer)
 })
 
 export default persistReducer(rootPersistConfig, rootReducer)
