+++ src/Routes/MatchInfo/matchInfo.scss
@@ -6,6 +6,7 @@
   width: 100%;
   height: 100%;
   position: relative;
+  overflow: hidden;
   z-index: 0;
   background-color: var(--e-global-color-secondary);
 
@@ -70,6 +71,10 @@
   &__teammates {
     width: 100%;
     margin: 1rem 0 1.5rem 0;
+  }
+
+  &__parallax {
+    @include margin-top-sm;
   }
 
   @include media-breakpoint-down(lg) {
