+++ src/components/Sections/ParallaxHeading/parallaxHeading.scss
